Extract canvas/webgl code to separate component
