+++ frontend/src/app/components/block-overview-graph/block-overview-graph.component.html
@@ -1,0 +1,6 @@
+<div class="block-overview-graph">
+  <canvas class="block-overview-canvas" #blockCanvas></canvas>
+  <div class="loader-wrapper" [class.hidden]="!isLoading">
+    <div class="spinner-border ml-3 loading" role="status"></div>
+  </div>
+</div>
