+++ frontend/src/app/components/block-overview-graph/block-overview-graph.component.scss
@@ -1,4 +1,4 @@
-.mempool-block-overview {
+.block-overview-graph {
   position: relative;
   width: 100%;
   padding-bottom: 100%;
@@ -8,12 +8,15 @@
   align-items: center;
 }
 
-.block-overview {
+
+.block-overview-canvas {
   position: absolute;
   left: 0;
   right: 0;
   top: 0;
   bottom: 0;
+  width: 100%;
+  height: 100%;
   overflow: hidden;
 }
 
@@ -27,6 +30,7 @@
   justify-content: center;
   align-items: center;
   transition: opacity 500ms 500ms;
+  pointer-events: none;
 
   &.hidden {
     opacity: 0;
