+++ frontend/src/app/components/block-overview-graph/fast-vertex-array.ts
