+++ frontend/src/app/components/block-overview-graph/sprite-types.ts
