+++ frontend/src/app/components/block-overview-graph/tx-sprite.ts
