+++ frontend/src/app/components/block-overview-graph/tx-view.ts
