+++ frontend/src/app/components/mempool-block-overview/mempool-block-overview.component.html
@@ -1,6 +1,7 @@
-<div class="mempool-block-overview">
-  <canvas class="block-overview" [style.width]="cssWidth + 'px'" [style.height]="cssHeight + 'px'" #blockCanvas></canvas>
-  <div class="loader-wrapper" [class.hidden]="!(isLoading$ | async)">
-    <div class="spinner-border ml-3 loading" role="status"></div>
-  </div>
-</div>
+<app-block-overview-graph
+  #blockGraph
+  [isLoading]="isLoading$ | async"
+  [resolution]="75"
+  [blockLimit]="stateService.blockVSize"
+  (txPreviewEvent)="onTxPreview($event)">
+</app-block-overview-graph>
