+++ frontend/src/app/shared/shared.module.ts
@@ -45,6 +45,7 @@ import { StartComponent } from '../components/start/start.component';
 import { TransactionComponent } from '../components/transaction/transaction.component';
 import { TransactionsListComponent } from '../components/transactions-list/transactions-list.component';
 import { BlockComponent } from '../components/block/block.component';
+import { BlockOverviewGraphComponent } from '../components/block-overview-graph/block-overview-graph.component';
 import { AddressComponent } from '../components/address/address.component';
 import { SearchFormComponent } from '../components/search-form/search-form.component';
 import { AddressLabelsComponent } from '../components/address-labels/address-labels.component';
@@ -110,6 +111,7 @@ import { SvgImagesComponent } from '../components/svg-images/svg-images.componen
     StartComponent,
     TransactionComponent,
     BlockComponent,
+    BlockOverviewGraphComponent,
     TransactionsListComponent,
     AddressComponent,
     SearchFormComponent,
@@ -203,6 +205,7 @@ import { SvgImagesComponent } from '../components/svg-images/svg-images.componen
     StartComponent,
     TransactionComponent,
     BlockComponent,
+    BlockOverviewGraphComponent,
     TransactionsListComponent,
     AddressComponent,
     SearchFormComponent,
